refactor(agents): isolate edit replacement reconstruction
